requirements.txt
discord.py==1.5.1
python-dotenv==0.14.0
aiomysql==0.0.20